pyproject.toml
[project]
name = "shop-catalog-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "langchain>=0.2.17",
    "langchain-community==0.2.16",
    "langchain-core>=0.2.43",
    "langchain-google-genai<=3.0.0",
    "langchain-pinecone==0.1.2",
    "mysql-connector-python>=9.5.0",
    "pandas>=2.3.3",
    "pinecone<=7.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
